UX Update, removed landing page and labels during imports
