+++ packages/nextjs/app/(home)/_components/TotalAmountDisplay.tsx
